Better error handling
This is an upstream sync of better error messaging. Now Synapse will detect if there are no applicable extensions for a given type (which it sort of did before but without much details) and give a printout of the environment. Similarly, if it has applicable extensions but no applicable models, it will error with more details as well. This avoids the obscure "could not indent -1" error that would otherwise emanate from JavaPoet.